requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
datasets==2.14.7
dill==0.3.7
python-dotenv==1.0.0